Overhaul statistics calculation
To provide more comprehensive info, we move statistics into a separate module (stats).
A summarizer is now fed individual rounds and generates round summaries for each. Once
the battle is finished, it can also produce a battle summary. Individual statistics are
tracked via a Stat object, which allows easy cumulative computation of the mean and
variance.