Surpress unfixable build warnings
On build, Angular throws two warnings - a circular dependency warning, and a
CommonJS warning. The circular dependency warning originates from wasm, and I
am unsure how to fix it, or if we even need to fix it. Unfortunately, there's
no way to ignore this specific warning, so unless I can figure out how to fix
this circular dependency, we won't be warned about *any* circular denpendency.

The other warning is for our charting library - it uses a CommonJS dependency,
meaning it can't be optimized effectively (and indeed, it bloats the size of
the app by quite a bit). This can't be fixed unless we switch to another
charting library, which we may do in the future. In the meantime, we just add
it to the list of allowed CommonJs dependencies.